+++ app/src/androidTest/java/com/jraska/github/client/xpush/PushServerDto.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.push
+package com.jraska.github.client.xpush
 
 import com.google.gson.annotations.SerializedName
